change test structure
